Add team api and support multiple apis
This introduces a new team api. I want to separate the transaction endpoints from the new team endpoints, and therefore have introduced a new concept called "api". An api is a group of endpoints. This transforms "the Signhost api" essentially into a "group of apis, of which the transaction api and the team api are currently part".
Visually and logically I tried to separate them without changing too much of the current design. I had to try a few different approaches and failed some. The current design is roughly satisfactory for now. Things that didn't work:
1. adding a new key X-Group property to an endpoint in the swagger api and then group them in liquid template
2. Adding the endpoints to a static group by means of adding them to a multidimensional array
3. matching endpoints based on the endpoint string

At some point I realized I'd just better simply use the swagger/openapi file format and separate the apis into their own file. As a consequence of separating the examples and swagger file, a subdirectory had to be introduced. So the solution looks like:

- _data
    - apis
        - transaction
            - swagger.yaml
            - examples.yaml
        - team
            - swagger.yaml
            - examples.yaml

Each subfolder of `apis` is treated as a separate api. `endpoints.html` now supports this. This solution simplifies management (as each api has their own folder + files). Furthermore, this allows simpler refactoring into multiple "service apis" in the future.